% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm19v.R
\name{q19}
\alias{q19}
\title{19v Normalization Function}
\usage{
q19(X)
}
\arguments{
\item{Data}{matrix UnImputed gene expression data, which represent gene expressions for each cell/sample. rows = genes and cols= cells.}
}
\description{
The main goal of this function is normalize the scRNA-seq data
}
\examples{
Norm19V ( Data )
}
\keyword{Norm19V}
