% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZeroModelThresholding.R
\name{ZeroModelThresholding}
\alias{ZeroModelThresholding}
\title{Zero Model thresholding Function}
\usage{
ZeroModelThresholding(
  Raw,
  imputed_scaled,
  clusters,
  zeroModel = "Unknown",
  size0 = -1,
  mu0 = -1,
  buffer = 0.05,
  skew = 1
)
}
\arguments{
\item{Raw}{data matrix UnImputed gene expression data, which represent gene expressions for each cell/sample. rows = genes and cols= cells.}

\item{Imputed}{data matrix WNMF/sNMF gene expression data, which represent gene expressions for each cell/sample. rows = genes and cols= cells.}
}
\description{
The main goal of this function is to apply a thresholding step to avoid over imputation
}
\examples{
ZeroModelThresholding ( Raw , scaledWNMF, clusters, , zeroModel = "Known", size0 = 1, mu0 = 50, buffer = 0)
}
\keyword{imputation}
